really adapt c flags for C++14 (used in the last source changes here & in APT)
CONFIG += c++14 # https://stackoverflow.com/a/27458874/94687

It's needed to accompany the change in
7691ee438d3b33bb9ed752c05e33e59e51f7cf8f with this change in *.pro.

This is not important for gcc of large enough versions (as in
Sisyphus), because the default language there includes C++14 (and no
additional flags are passed by default). However, this is needed for
lcc (gcc substitute on e2k), because its default language is not like
this, and perhaps not even C++11, so normally, qmake passes
-std=gnu++11 flag to it when building Qt apps (and so we need to
override it here).