checker: compile as C++11 to ensure the code is ready for it
Here it is done for p8, similarly to how it was done
(with CONFIG += c++11) in:

25cba70 really adapt c flags for C++14 (used in the last source changes here & in APT)

This is not important for gcc of large enough versions (as in
Sisyphus), because the default language there includes C++14 (and no
additional flags are passed by default). However, this is needed for
lcc (gcc substitute on e2k), because its default language is not like
this, and perhaps not even C++11, so normally, qmake passes
-std=gnu++11 flag to it when building Qt apps (and so we need to
override it here).