checker: -Woverloaded-virtual made an error (to avoid APT API misuse)
This helps prevent coding errors when an override of a method was
intended, but it has a wrong signature: then it would trigger this
warning, because the overriding method in the derived class hides all
other overloaded methods from the base class (according to the C++
standard).

Unfortunately, lcc in p8_e2k (1.21.24) doesn't know this flag.
So, we exclude it for the e2k archs according to
<https://stackoverflow.com/q/356666/94687>. (Note that
QMAKE_TARGET.arch didn't work for us, whereas QMAKE_HOST.arch did work.)