Revert "Update to new Apt ABI" -- to follow the order of the upcoming patches in APT
Some of those upcoming patches also affect ABI, so we'd like to be
able to build intermediate revisions of APT and apt-indicator and test them.

This reverted change will be applied at the end of the current patch
series. Or maybe not, because it brings just a dim advantage in speed
by passing const arguments by reference whereas brings the
disadvantage in being incompatible with the upstream of other clients
of APT, which use the current Debian's APT API (without this change);
so, it adds a bit more burden on the maintainer for little gain.

This reverts commit 7691ee438d3b33bb9ed752c05e33e59e51f7cf8f.

But keep all override marks on the methods from past commits.)